feat: added a footer with content from filter drawer that is more appropriate in a footer
